Attempt to resolve index miscalculation
After having fixed the coordinate system to a more coherent one the index system stopped working. Created formula R * Y + X where R is the number of rows (or the width) of the map to identify the index of selected hexagon. The raycast appears to not be working, or maybe it's the conversion from real world coordinates to offset coordinates.